update Ocelloids implementation w hook
